pyproject.toml
[build-system]
requires = [
    "setuptools >= 61.0",
] # add any imports required by setup.py here
build-backend = "setuptools.build_meta"

# for more info see https://packaging.python.org/en/latest/guides/writing-pyproject-toml/

[project]
name = "animodel"
authors = [{ name = "Tomas Jakab", email = "tomj@robots.ox.ac.uk" }]
dynamic = ["version"]                                                # let setup.py set version
requires-python = ">= 3.8"
dependencies = [
    "numpy",
    "pandas",
    "trimesh",
    "tqdm",
    "scipy",
    "hydra-core",
    "omegaconf",
    "scikit-image",
]

[project.optional-dependencies]
# add any here
test = []

[project.urls]
repo = "https://github.com/tomasjakab/animodel"
